+++ src/main/java/gui/content/ContentController.java
@@ -1,6 +1,4 @@
-package gui;
-
-import java.io.IOException;
+package gui.content;
 
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -10,7 +8,10 @@ import javafx.scene.control.TextField;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.AnchorPane;
 
-public class HomePageController {
+import java.io.IOException;
+
+public class ContentController {
+
     @FXML
     private ImageView image1;
 
@@ -59,6 +60,7 @@ public class HomePageController {
 
     @FXML
     void loadSearchPage(ActionEvent event) throws IOException {
+        //System.out.println("search bar function");
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/searchpage/SearchPage.fxml"));
