+++ src/main/java/gui/episodeoverview/EpisodeOverviewController.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.episodeoverview;
 
 import java.io.IOException;
 
@@ -10,7 +10,7 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 
-public class SeasonOverviewController {
+public class EpisodeOverviewController {
 
     @FXML
     private Button homeButton;
@@ -26,31 +26,33 @@ public class SeasonOverviewController {
 
     @FXML
     private MenuItem seasonSelector;
-
+    
     @FXML 
     private MenuItem episodeSelector;
 
     @FXML
     void loadHomePage(ActionEvent event) throws IOException {
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/homepage/HomePage.fxml"));
+
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
 
     @FXML
     void loadSeasonPage(ActionEvent event) throws IOException {
-
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/seasonoverview/SeasonOverview.fxml"));
+
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
 
     @FXML
     void loadEpisodePage(ActionEvent event) throws IOException{
-
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/episodeoverview/EpisodeOverview.fxml"));
+
 
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
+
 }
