+++ src/main/java/gui/showoverview/ShowOverviewController.java
@@ -1,4 +1,4 @@
-package gui;
+package gui.showoverview;
 
 import java.io.IOException;
 
@@ -10,7 +10,7 @@ import javafx.scene.control.MenuItem;
 import javafx.scene.control.TextField;
 import javafx.scene.layout.AnchorPane;
 
-public class EpisodeOverviewController {
+public class ShowOverviewController {
 
     @FXML
     private Button homeButton;
@@ -22,25 +22,25 @@ public class EpisodeOverviewController {
     private TextField searchBar;
 
     @FXML
-    private TextField searchBar1;
+    private MenuItem seasonSelector;
 
     @FXML
-    private MenuItem seasonSelector;
-    
-    @FXML 
     private MenuItem episodeSelector;
 
-    @FXML
+     @FXML
     void loadHomePage(ActionEvent event) throws IOException {
-        FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/homepage/HomePage.fxml"));
 
-        AnchorPane pane = loader.load();
+         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/homepage/HomePage.fxml"));
+
+         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
 
     @FXML
     void loadSeasonPage(ActionEvent event) throws IOException {
+
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/seasonoverview/SeasonOverview.fxml"));
+
 
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
@@ -50,9 +50,7 @@ public class EpisodeOverviewController {
     void loadEpisodePage(ActionEvent event) throws IOException{
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fxml/episodeoverview/EpisodeOverview.fxml"));
 
-
         AnchorPane pane = loader.load();
         rootPane.getChildren().setAll(pane);
     }
-
 }
