+++ src/main/java/module-info.java
@@ -12,4 +12,18 @@ module Application {
     exports edu.metrostate;
    // exports gui.src;
     exports gui;
+    exports gui.content;
+    opens gui.content to javafx.fxml;
+    exports gui.episodeoverview;
+    opens gui.episodeoverview to javafx.fxml;
+    opens gui.homepage to javafx.fxml;
+    exports gui.homepage;
+    exports gui.login;
+    opens gui.login to javafx.fxml;
+    exports gui.searchpage;
+    opens gui.searchpage to javafx.fxml;
+    exports gui.seasonoverview;
+    opens gui.seasonoverview to javafx.fxml;
+    exports gui.showoverview;
+    opens gui.showoverview to javafx.fxml;
 }
