+++ src/main/resources/fxml/content/Content.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <BorderPane xmlns="http://javafx.com/javafx/8.0.65"
             xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="gui.ContentController">
+            fx:controller="gui.content.ContentController">
 
 
     <!-- Top contains search bar, home button and login button-->
